+++ src/api/hooks/calendar.ts
@@ -1,6 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
 
-import { CalendarResponse } from '@/api/types/calendar'
+import { CalendarResponse, GetCalendarYearlyRequest, GetCalendarYearlyResponse } from '@/api/types/calendar'
+import { getInitialAcademicCalendar } from '@/util/academicCalendar'
 import { apiInterface } from '@/util/axios/custom-axios'
 
 const getCalendar = async (year: number, month: number) => {
@@ -13,5 +14,18 @@ export const useGetCalendar = (year: number, month: number) => {
     queryKey: ['calendar', year, month],
     queryFn: () => getCalendar(year, month),
     initialData: [{ date: new Date(), event: [], eventCount: 0 }] as CalendarResponse[],
+  })
+}
+
+const getAcademicCalendar = async (params: GetCalendarYearlyRequest) => {
+  const response = await apiInterface.get<GetCalendarYearlyResponse>('/calendar/academic', { params })
+  return response.data
+}
+
+export const useGetAcademicCalendar = ({ year, semester }: GetCalendarYearlyRequest) => {
+  return useQuery({
+    queryKey: ['academic-calendar', year, semester],
+    queryFn: () => getAcademicCalendar({ year, semester }),
+    initialData: getInitialAcademicCalendar(semester),
   })
 }
