+++ src/api/types/calendar.ts
@@ -1,3 +1,5 @@
+import { MonthEventType } from '@/types/calendar'
+
 export interface CalendarEventProps {
   id: number
   title: string
@@ -8,3 +10,10 @@ export interface CalendarResponse {
   event: CalendarEventProps[]
   eventCount: number
 }
+
+export interface GetCalendarYearlyRequest {
+  year: number
+  semester: 1 | 2
+}
+
+export type GetCalendarYearlyResponse = MonthEventType[]
