--- src/api/types/home_sub.ts
@@ -1,8 +1,0 @@
-import { MonthEventType } from '@/types/home_sub'
-
-export interface GetCalendarYearlyRequest {
-  year: number
-  semester: 1 | 2
-}
-
-export type GetCalendarYearlyResponse = MonthEventType[]
