+++ src/types/calendar.ts
@@ -4,3 +4,17 @@ export interface DayProps {
   date: Date
   selected: boolean
 }
+
+export interface EventType {
+  id: number
+  title: string
+  description: string
+  startDate: string
+  startDay: string
+  endDate: string
+  endDay: string
+}
+export interface MonthEventType {
+  month: number
+  schedules: EventType[]
+}
